Refactor: Simplify calculation of external vs. internal modules.
Ensure that all internal modules receive a fully qualified name. External modules are identified by having module paths that do not start with the path of the base module used for scanning (they can however start with the path of the root module, if the base module is not the root module. This means that sibling modules of the base module are considered external).