Implement aliases for to simplify plot labels.
It is now possible to specify aliases in the visualize
method, to shorten plot labels. This is helpful when a
module has many submodules and/or a long name that would
crowd the plot.
To avoid a runtime error where python 3.8 cannot be found
when running the pre commit hooks, the language version
was set to 3.10, which is the version that is installed
when setting up the project.
A discussion on the error can be found here:
https://github.com/pre-commit/pre-commit/issues/1375#issue-587595366