Feature: Allow specification of modules by only part of their names.
With a syntax of (*)X(*), modules will be retrieved that contain X in their name (and start with X, if no preceding */end with X, if no * afterwards). The list of modules in the graph will be traversed to match them against the user-specified partial names. If they match, the modules will be considered instead of their partial representatives. Multiple modules can match one partial representative.
If no modules match a partial representative, an error will be raised to alert the user to their mis-specification.