Feature: Ensure that no submodule is cut off during the regex matching.
This could otherwise throw off the generation of error messages, if modules that do match a regex cannot be associated with the regex, if they are submodules of other matching modules. Add warning to the documentation to alert the user to the possible runtime effect of using a non-restrictive regex pattern.